requirements.txt
allure-python-commons
pytest
selenium
requests
allure-pytest
pytest-xdist
webdriver-manager
